refactor: different method of handling the callback
